feat: add seata-order Service

+++ seata-account-service/src/main/java/com/example/seataaccountservice/controller/AccountController.java
@@ -7,8 +7,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigDecimal;
-
 /**
  * @author Real
  * @since 2023/6/4 0:06
@@ -29,7 +27,7 @@ public class AccountController {
      * @return {@link CommonResult}
      */
     @RequestMapping("/decrease")
-    public CommonResult decrease(@RequestParam("userId") Long userId, @RequestParam("money") BigDecimal money) {
+    public CommonResult decrease(@RequestParam("userId") Long userId, @RequestParam("money") Long money) {
         accountService.decrease(userId, money);
         return new CommonResult("扣减账户余额成功！", 200);
     }
