+++ seata-account-service/src/main/java/com/example/seataaccountservice/mapper/AccountMapper.java
@@ -20,7 +20,7 @@ public interface AccountMapper extends BaseMapper<Account> {
      * @param userId 用户id
      * @param money  金额
      */
-    void decrease(@Param("userId") Long userId, @Param("money") BigDecimal money);
+    void decrease(@Param("userId") Long userId, @Param("money") Long money);
 
 }
